+++ dio_cache_interceptor/lib/src/extension/request_extension.dart
@@ -3,17 +3,6 @@ import 'package:dio_cache_interceptor/src/extension/cache_option_extension.dart'
 import 'package:http_cache_core/http_cache_core.dart';
 
 extension RequestExtension on RequestOptions {
-  // Request headers are not of type Headers...
-  // This method helps to wrap it to correct list for iteration
-  List<String>? headerValuesAsList(String headerKey) {
-    final value = headers[headerKey];
-
-    if (value is Iterable<String>) return value.toList();
-    if (value is String) return value.split(',').map((h) => h.trim()).toList();
-
-    return value;
-  }
-
   CacheOptions? getCacheOptions() {
     return extra[extraKey];
   }
@@ -24,7 +13,7 @@ extension RequestExtension on RequestOptions {
 
     for (var header in headers.entries) {
       if (header.value is Iterable) {
-        h[header.key] = header.value.join(',');
+        h[header.key] = header.value.join(', ');
       } else if (header.value != null) {
         h[header.key] = header.value.toString();
       }