+++ dio_cache_interceptor/lib/src/model/dio_base_request.dart
@@ -5,12 +5,10 @@ import 'package:http_cache_core/http_cache_core.dart';
 /// Wrapper on [BaseRequest] to deal with [RequestOptions]
 class DioBaseRequest extends BaseRequest {
   final RequestOptions request;
+  final Map<String, String> _headers;
 
-  DioBaseRequest(this.request);
-
-  @override
-  List<String>? headerValuesAsList(String header) {
-    return request.headerValuesAsList(header);
+  DioBaseRequest(this.request) : _headers = request.getFlattenHeaders() {
+    request.headers = _headers;
   }
 
   @override
@@ -21,4 +19,7 @@ class DioBaseRequest extends BaseRequest {
       request.headers[header] = value;
     }
   }
+
+  @override
+  Map<String, String> get headers => _headers;
 }