+++ http_cache_core/lib/src/model/core/base_request.dart
@@ -1,7 +1,7 @@
 /// Abtract request adapter
 abstract class BaseRequest {
-  /// Header values as list for a given header key.
-  List<String>? headerValuesAsList(String header);
+  /// Headers.
+  Map<String, String> get headers;
 
   /// Applies a [value] if not null to the request's header or remove it otherwise.
   void setHeader(String header, String? value);