+++ http_cache_isar_store/lib/src/store/http_cache_isar_store.dart
@@ -178,7 +178,7 @@ class IsarCacheStore extends CacheStore {
 
   CacheResponse _mapDataToResponse(Cache data) {
     return CacheResponse(
-      cacheControl: CacheControl.fromHeader(data.cacheControl?.split(', ')),
+      cacheControl: CacheControl.fromString(data.cacheControl),
       content: data.content,
       date: data.date,
       eTag: data.eTag,